version 0.3a:
  - brand new module structure;
  - ability to disable some fields and versions for sync;
  - and much more.
