+++ app/views/loader/new.html.erb
@@ -1,18 +1,18 @@
-<%= stylesheet_link_tag 'loader', :plugin => 'redmine_loader' %>
+<%= stylesheet_link_tag 'loader', plugin: 'redmine_loader' %>
 <div class='left half'>
   <h1><%= l(:import_tasks) %></h1>
-  <%= form_for(:import, :url => analyze_project_loader_path, :html => {:id => 'import', :multipart => true}) do |f| %>
-    <%= render :partial => 'about' %>
+  <%= form_for(:import, url: analyze_project_loader_path, html: {id: 'import', multipart: true}) do |f| %>
+    <%= render partial: 'about' %>
     <%= f.file_field 'xmlfile' %>
     <tr>
-      <%= f.submit(l(:analyze), {:name => 'import[analyze]', :disable_with => l(:analyzing)}) %>
+      <%= f.submit(l(:analyze), {name: 'import[analyze]', disable_with: l(:analyzing)}) %>
     </tr>
   <% end %>
 </div>
 <div class='right half'>
   <h1><%= l(:export_tasks) %></h1>
-  <%= form_for(:import, {:url => export_project_loader_path, :method => :get, :html => {:id => 'export', :multipart => true}}) do |f| %>
+  <%= form_for(:import, {url: export_project_loader_path, method: :get, html: {id: 'export', multipart: true}}) do |f| %>
     <p><%= l(:generate_xml) %></p>
-    <%= f.submit(l(:export_issue), {:name => 'export'}) %>
+    <%= f.submit(l(:export_issue), {name: 'export'}) %>
   <% end %>
 </div>
