+++ app/views/settings/_loader_settings.html.erb
@@ -1,17 +1,70 @@
-<% html_title "Exony Project File Loader" -%>
-<%= content_tag :p do %>
-  <%= content_tag :label, 'Mark issues as private by default' %>
-  <%= check_box_tag "settings[is_private_by_default]", 1, Setting[:plugin_redmine_loader][:is_private_by_default] %>
-<% end %>
-<%= content_tag :p do %>
-  <%= content_tag :label, 'Alias for Tracker field' %>
-  <%= text_field_tag 'settings[tracker_alias]', @settings['tracker_alias'] %>
-<% end %>
-<%= content_tag :p do %>
-  <%= content_tag :label, 'Max number of tasks to immediately import' %>
-  <%= number_field_tag 'settings[instant_import_tasks]', @settings['instant_import_tasks'], :in => 10..100, :step => 5 %>
-<% end %>
-<%= content_tag :p do %>
-  <%= content_tag :label, 'Tracker for project files' %>
-  <%= select_tag "settings[tracker_id]", options_from_collection_for_select(Tracker.all, :id, :name, :selected => @settings['tracker_id']) %>
-<% end %>
+<fieldset class="box" id="import_ignore_fields"><legend>Import settings</legend>
+  <%= content_tag :p do %>
+    <%= content_tag :label, 'Mark issues as private by default' %>
+    <%= check_box_tag "settings[import][is_private_by_default]", 1, @settings[:import][:is_private_by_default] %>
+  <% end %>
+  <%= content_tag :p do %>
+    <%= content_tag :label, 'Sync versions' %>
+    <%= check_box_tag "settings[import][sync_versions]", 1, @settings[:import][:sync_versions] %>
+  <% end %>
+  <%= content_tag :p do %>
+    <%= content_tag :label, 'Alias for Redmine issue ID field' %>
+    <%= text_field_tag 'settings[import][redmine_id_alias]', @settings[:import][:redmine_id_alias] %>
+  <% end %>
+  <%= content_tag :p do %>
+    <%= content_tag :label, 'Alias for Tracker field' %>
+    <%= text_field_tag 'settings[import][tracker_alias]', @settings[:import][:tracker_alias] %>
+  <% end %>
+  <%= content_tag :p do %>
+    <%= content_tag :label, 'Max number of tasks to immediately import' %>
+    <%= number_field_tag 'settings[import][instant_import_tasks]', @settings[:import][:instant_import_tasks], in: 10..100, step: 5 %>
+  <% end %>
+  <%= content_tag :p do %>
+    <%= content_tag :label, 'Tracker for project files' %>
+    <%= select_tag "settings[import][tracker_id]", options_from_collection_for_select(Tracker.all, :id, :name, selected: @settings[:import][:tracker_id]) %>
+  <% end %>
+  <fieldset class="box" id="import_ignore_fields"><legend>Ignore fields</legend>
+    <%= content_tag :p do %>
+      <%= content_tag :label, 'Estimated hours' %>
+      <%= check_box_tag "settings[import][ignore_fields][estimated_hours]", 1, @settings[:import][:ignore_fields][:estimated_hours] %>
+    <% end %>
+    <%= content_tag :p do %>
+      <%= content_tag :label, 'Priority' %>
+      <%= check_box_tag "settings[import][ignore_fields][priority_id]", 1, @settings[:import][:ignore_fields][:priority_id] %>
+    <% end %>
+    <%= content_tag :p do %>
+      <%= content_tag :label, 'Description' %>
+      <%= check_box_tag "settings[import][ignore_fields][description]", 1, @settings[:import][:ignore_fields][:description] %>
+    <% end %>
+    <%= content_tag :p do %>
+      <%= content_tag :label, 'Done ratio' %>
+      <%= check_box_tag "settings[import][ignore_fields][done_ratio]", 1, @settings[:import][:ignore_fields][:done_ratio] %>
+    <% end %>
+  </fieldset>
+</fieldset>
+
+
+<fieldset class="box" id="export_ignore_fields"><legend>Export settings</legend>
+  <%= content_tag :p do %>
+    <%= content_tag :label, 'Sync versions' %>
+    <%= check_box_tag "settings[export][sync_versions]", 1, @settings[:export][:sync_versions] %>
+  <% end %>
+  <fieldset class="box" id="export_ignore_fields"><legend>Ignore fields</legend>
+    <%= content_tag :p do %>
+      <%= content_tag :label, 'Estimated hours' %>
+      <%= check_box_tag "settings[export][ignore_fields][estimated_hours]", 1, @settings[:export][:ignore_fields][:estimated_hours] %>
+    <% end %>
+    <%= content_tag :p do %>
+      <%= content_tag :label, 'Priority' %>
+      <%= check_box_tag "settings[export][ignore_fields][priority_id]", 1, @settings[:export][:ignore_fields][:priority_id] %>
+    <% end %>
+    <%= content_tag :p do %>
+      <%= content_tag :label, 'Description' %>
+      <%= check_box_tag "settings[export][ignore_fields][description]", 1, @settings[:export][:ignore_fields][:description] %>
+    <% end %>
+    <%= content_tag :p do %>
+      <%= content_tag :label, 'Done ratio' %>
+      <%= check_box_tag "settings[export][ignore_fields][done_ratio]", 1, @settings[:export][:ignore_fields][:done_ratio] %>
+    <% end %>
+  </fieldset>
+</fieldset>
