+++ lib/import_task.rb
@@ -2,22 +2,22 @@ class ImportTask < Struct.new(:assigned_to,
                         :category,
                         :code,
                         :delays,
-                        :duration,
-                        :finish,
+                        :estimated_hours,
+                        :due_date,
                         :is_private,
                         :level,
                         :milestone,
-                        :notes,
+                        :description,
                         :outlinenumber,
                         :outlinelevel,
                         :parent_id,
-                        #:percentcomplete,
+                        :done_ratio,
                         :predecessors,
-                        #:priority,
-                        :start,
+                        :priority_id,
+                        :start_date,
                         :status_id,
                         :tid,
-                        :title,
+                        :subject,
                         :tracker_id,
                         :tracker_name,
                         :uid)
